requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
phonenumbers==9.0.10
pillow==11.3.0
propcache==0.3.2
psycopg2==2.9.10
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
twilio==9.7.0
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
